Mysql: use correct log file and position
According to several experts (including Percona), when issuing "CHANGE
MASTER TO" you are specifying the position in the master binary log that
the slave will start executing from.

Exec_Master_Log_Pos and Relay_Master_Log_File are the correct positions
to save when reloading this snapshot and issuing a "CHANGE MASTER TO"
command.

Percona:
http://www.mysqlperformanceblog.com/2008/07/07/how-show-slave-status-relates-to-change-master-to/

Pythian:
http://www.pythian.com/blog/beware-starting-slaves-in-position-in-the-master-info-file/