Make usage more convenient by checking extra paths for fsfreeze.
`ec2-consistent-snapshot` can succeed in manual testing but then fail
when run via cron because cron sets a more restrictive $PATH variable.

This came up earlier this year in related bug report:
https://github.com/alestic/ec2-consistent-snapshot/issues/49

Some other folks have blogged about the gotcha in the past as well.

The issue notably affects "fsfreeze" which is in /sbin on Ubuntu
and /usr/sbin on Fedora, and is not found by cron by default.

People can workaround the issue by setting a PATH variable, but this
patch provides a better user experience by checking in /sbin and
/usr/sbin for fsfreeze even if they are not in $PATH.