Downgrade debian/compat to 10
The debhelper class does not work with compat level 11, because it tries to
call dh_systemd_enable, which has been deprecated, and errors out in that
compat level.

(https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=885407)

As we're not providing systemd units at all, simply downgrading to compat
level 10 seems to be a safe solution.